move top-level info to top-level state
