+++ views/dashboard.js
@@ -2,20 +2,44 @@
 'use strict';
 
 var React = require('react'),
-  _ = require('lodash'),
+  $ = require('jquery'),
   ChartElement = require('./chart-element'),
-  url = require('url');
+  url = require('url'),
+  org = url.parse(window.location.href).pathname.split('/')[1];
 
 var Dashboard = React.createClass({
   getInitialState: function () {
     return {
-
+      loaded: false,
+      schema: {}
     };
   },
 
-  //componentDidMount: function () {
+  componentDidMount: function () {
 
-  //},
+    // make sure that we're logged in
+    $.ajax({
+      url: '/' + org + '/browser-api/v1/resources/honorific?count=true',
+      dataType: "json",
+      success: function (data) {
+        // nothing to do
+      }.bind(this),
+      error: function (err) {
+        // we're probably not logged in to the server
+        window.location = '/' + org + '/logout';
+      }.bind(this)
+    });
+
+    // fetch the discovery document
+    $.ajax({
+      url: '/' + org + '/discovery/v1alpha1/apis/v1alpha1/rest',
+      dataType: "json",
+      success: function (data) {
+        this.setState({schema: data, loaded: true});
+      }.bind(this)
+    });
+
+  },
 
   render: function () {
 
@@ -24,18 +48,30 @@ var Dashboard = React.createClass({
         <div className="panel panel-info">
           <div>
             <div className="col-md-6">
-              <ChartElement position={0}></ChartElement>
+              <ChartElement
+                loaded={this.state.loaded}
+                schema={this.state.schema}
+                position={0}></ChartElement>
             </div>
             <div className="col-md-6">
-              <ChartElement position={1}></ChartElement>
+              <ChartElement
+                loaded={this.state.loaded}
+                schema={this.state.schema}
+                position={1}></ChartElement>
             </div>
           </div>
           <div>
             <div className="col-md-6">
-              <ChartElement position={2}></ChartElement>
+              <ChartElement
+                loaded={this.state.loaded}
+                schema={this.state.schema}
+                position={2}></ChartElement>
             </div>
             <div className="col-md-6">
-              <ChartElement position={3}></ChartElement>
+              <ChartElement
+                loaded={this.state.loaded}
+                schema={this.state.schema}
+                position={3}></ChartElement>
             </div>
           </div>
         </div>
